Move context for field, file, featured-item, dropdown, dialog
