+++ src/generic/generic-component-dropdown/data/demo.js
@@ -1,0 +1,22 @@
+module.exports = {
+  _demo: {
+    scripts: `
+      document.addEventListener('DOMContentLoaded', function () { ECL.initExpandables(); });
+      ECL.dropdown('.ecl-dropdown');
+      `,
+  },
+  links: [
+    {
+      target: '#',
+      title: 'Article',
+    },
+    {
+      target: '#',
+      title: 'Page',
+    },
+    {
+      target: '#',
+      title: 'Community',
+    },
+  ],
+};
