+++ src/generic/generic-component-featured-item/data/demo--default.js
@@ -1,0 +1,12 @@
+module.exports = {
+  image: {
+    src: '../../assets/demo_images/technics-demo-1.jpg',
+    alt: 'Technic picture',
+  },
+  link: {
+    href: '../../example.html#featured-item-1',
+    label: 'See more',
+  },
+  metas: ['announcement', '21 December 2016'],
+  title: 'Energy Union tour',
+};
