+++ src/generic/generic-component-featured-item/data/demo--extended.js
@@ -1,0 +1,13 @@
+module.exports = {
+  variant: 'extended',
+  image: {
+    src: '../../assets/demo_images/technics-demo-1.jpg',
+    alt: 'Technic picture',
+  },
+  metas: ['announcement', '21 December 2016'],
+  title: 'Energy Union tour',
+  link: {
+    href: '../../example.html#featured-item-2',
+    label: 'See more',
+  },
+};
