+++ src/systems/ec/ec-component/ec-component-dropdown/ec-component-dropdown.config.js
@@ -1,29 +1,10 @@
+const context = require('@ecl/generic-component-dropdown/data/demo');
+
 module.exports = {
   title: 'Dropdowns',
   label: 'Dropdowns',
   status: 'ready',
   preview: '@preview-center-transparent',
   tags: ['molecule'],
-  context: {
-    _demo: {
-      scripts: `
-        document.addEventListener('DOMContentLoaded', function () { ECL.initExpandables(); });
-        ECL.dropdown('.ecl-dropdown');
-        `,
-    },
-    links: [
-      {
-        target: '#',
-        title: 'Article',
-      },
-      {
-        target: '#',
-        title: 'Page',
-      },
-      {
-        target: '#',
-        title: 'Community',
-      },
-    ],
-  },
+  context,
 };
