+++ src/systems/eu/eu-component/eu-component-dropdown/package.json
@@ -13,7 +13,8 @@
     "@ecl/eu-component-expandable": "^0.0.1",
     "@ecl/eu-style-icon": "^0.0.1",
     "@ecl/eu-component-link-block": "^0.0.1",
-    "@ecl/eu-component-dropdown": "^0.0.1"
+    "@ecl/eu-component-dropdown": "^0.0.1",
+    "@ecl/generic-component-dropdown": "^0.0.1"
   },
   "publishConfig": { "access": "public" },
   "repository": {
